backend/requirements.txt
Django~=4.1.7
djangorestframework
markdown
django-filter
django-environ
postgres
psycopg2
Pillow
model_bakery